requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Base de Datos
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validación
pydantic==2.5.0
pydantic-settings==2.1.0

# Templates
jinja2==3.1.3

# Analytics y detección
user-agents==2.2.0
geoip2==4.7.0

# URLs cortas
shortuuid==1.0.11

# Pagos (opcional - comentado por ahora)
# stripe==7.10.0

# Redis (opcional - comentado por ahora)
# redis==5.0.1

# CORS
python-multipart==0.0.6
